fix Zero filter bug

+++ src/components/DepositBody.tsx
@@ -14,7 +14,7 @@ const DepositBody: React.FunctionComponent<{
   const [amount, setAmount] = useState<string>('');
 
   const amountGtBalance = utils.parseEther(amount || '0').gt(balance);
-  const amountLteZero = !amount || parseInt(amount) <= 0;
+  const amountLteZero = !amount || parseFloat(amount) <= 0;
 
 	return (
 		<div className="modal__deposit">
